Packaging fixes and improvements
Updates build and push scripts in line with other components.

This adds a "tyk" user and group for DEB and RPM packages so that package files and directories would be owned by it and the process run with its effective uid and gid. Appropriate changes made to build scripts and init configs.

The config file is now only readable and writable by the "tyk" user/group.

The build and push scripts allow manually assigning a version based on an environment variable, which is useful for unstable packaging (e.g. based on git revision ID).

Doesn't remove "default" init system files on upgrade/remove anymore as it's now a way for users to run the process with custom environment variables, hence sort of a config file.

Does a bunch of fixes on the packaging and init scripts (e.g. fixes RPM upgrades removing init scripts on upgrade). Removes old stuff. Adds Ubuntu 16.04 LTS (Xenial) package to PackageCloud.