Pump filtering (#212)
New feature to add the possibility of pump filtering.  This is related to issue #148 .

This adds a new config in each pump called `filters`  and it structure is the following:
 ```
"filters":{
        "api_ids":[],
        "org_ids":[],
        "skip_api_ids":[],
        "skip_org_ids":[]
      }
```
The fields `api_ids` and `org_ids` works as whitelists (APIs and orgs where we want to send the analytics records) and the fields  `skip_api_ids` and `skip_org_ids` works as the opposite (blackslits). 
The priority is always blacklisted configurations over whitelisted.

An example of configuration would be:
 ```
"csv2": {
      "type": "csv",
      "filters": {
        "org_ids": ["org1","org2"]
      },
      "meta": {
        "csv_dir": "./bar"
      }
    },
"csv3": {
      "type": "csv",
      "filters": {
        "org_ids":["org3"],
        "api_ids": ["api1"]
      },
      "meta": {
        "csv_dir": "./baz"
      }
    }
```