Add new filters based on response_code's (#270)
Add new filters based on response_code's

## Description
Currently we have the necessity to be able to only send with pump request with errors. For that we need to implement this new filter to be able to remove all 200 response code.


## Motivation and Context
Decrease space required to store all data of analytics . Request with 200 are not important to be monitorized and also are the 99% of traffic. We have some problems regarding sizing of our cluster of kafka and we want to reduce the load.

## How This Has Been Tested
-Unit test
-Start pump in local with same config of our staging environment with the new filter.