Sync automation workflows automatically (#300)
# Description
These related PRs,

Add a new workflow triggered on branch delete that will help with cleanup of developer environments. This workflow was forgotten in the last commit.

Add a workflow (sync-automation.yml) that will sync the automation workflows from master to known release branches. The list of branches is maintained [pr.zsh](https://github.com/TykTechnologies/tyk-ci/blob/master/wf-gen/pr.zsh). The workflow is only triggerred for appropriate paths and will only work for for the FILES in `pr.zsh`.

# Testing
Syntax checks have passed. Most of the real functionality is external to the workflow.

The status of the integration workflows can be inspected in this PR itself in the _Checks_ tab. The branch is prefixed with `feature/` to test that it triggers the required workflows when it is deleted.

The automation sync workflow was tested on [tyk-pump](https://github.com/TykTechnologies/tyk-pump/actions?query=workflow%3A%22Sync+automation%22) and the results can be viewed there.

# Merging
Once this PR is merged to master it will sync the automation workflows to all branches in the `matrix->branch` node.

And it will continue to do so for future commits.