Fix github release title (#323)
Sync by tyk-ci/wf-gen/pr.zsh. This commit was generated by code© and reviewed by humans™.

# Description
These related PRs fix the problem in https://github.com/TykTechnologies/tyk-analytics/runs/2115091960

It also adds `.goreleaser.yml` to the list of paths that are triggers for the automated sync that the _Merging_ section below depends on.

# Testing
The status of the workflows can be inspected in this PR itself in the _Checks_ tab. Since the github release function is only activated when a tag is pushed, it will be exercised when the next tag is pushed.

# Merging
Once this PR is merged, it will be automatically sync'd to release-3-lts and release-3.2.