[TT-2932] more control over upgrades (#384)
## Description
This PR is end result of a long chain of decisions. The immediate problem is that due to a bug in the release engineering code version 3.0.6 has a bug whereby the systemd service file is lost on upgrade. Since it is not possible to fix the releng bug in released packages, a workaround has been implemented.

This PR is dependant on the the install scripts being templates in tyk-ci/wf-gen. 

## Related Issue
https://tyktech.atlassian.net/browse/TT-3932

## How This Has Been Tested
See the status in the _Checks_ tab.

# Merging
This commit needs to be cherry-picked to release-3-lts, release-4 and any other active release branches. If the tykbot cherry pick fails, create a fresh PR instead of a rebase.

<!-- Automatic merging will be performed for the following branches when this PR is merged to master. -->

<!-- | Repo          | Branches                 | -->
<!-- |---------------|--------------------------| -->
<!-- | tyk           | release-3.2              | -->
<!-- | tyk-analytics | release-3.2              | -->
<!-- | tyk-pump      | release-0.8, release-1.0 | -->
<!-- | tyk-sink      | release-1.7, release-1.8 | -->

<!-- Please cherry pick this into the following branches: -->

<!-- | Repo          | Branches                 | -->
<!-- |---------------|--------------------------| -->
<!-- | tyk           | release-3, release-3-lts | -->
<!-- | tyk-analytics | release-3, release-3-lts | -->