chore: Update Dockerfile.std with GOEXPERIMENT environment variable
This commit updates the Dockerfile.std file by adding the GOEXPERIMENT environment variable. The variable is passed as an argument and set in the environment to enable experimental Go features. This change is made to enhance the build process and enable the use of experimental Go features in the CI environment.